+++ cloud-collect-core/src/main/kotlin/com/windea/demo/cloudcollect/core/configuration/AuthorizationServerConfiguration.kt
@@ -18,7 +18,7 @@ class AuthorizationServerConfiguration(
 	private val accessTokenConverter: JwtAccessTokenConverter,
 	private val authenticationManager: AuthenticationManager,
 	private val passwordEncoder: PasswordEncoder,
-	private val jwtProperties: JwtProperties
+	private val jwtProperties: SecurityProperties
 ) : AuthorizationServerConfigurerAdapter() {
 	override fun configure(configurer: ClientDetailsServiceConfigurer) {
 		configurer.inMemory()