+++ cloud-collect-core/src/main/kotlin/com/windea/demo/cloudcollect/core/properties/SecurityProperties.kt
@@ -3,10 +3,10 @@ package com.windea.demo.cloudcollect.core.properties
 import org.springframework.boot.context.properties.*
 import org.springframework.context.annotation.*
 
-/**Jwt的属性类。*/
+/**安全校验的属性类。*/
 @Configuration
-@ConfigurationProperties("com.windea.security.jwt")
-class JwtProperties {
+@ConfigurationProperties("com.windea.security")
+class SecurityProperties {
 	lateinit var signingKey: String
 	
 	lateinit var verifierKey: String